discussions: factor out mentions parsing to seperate package
